requirements.txt
Django==3.2.20
django-mysql==4.11.0
django-environ==0.11.2
mysqlclient==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
flex==6.14.1